COR: Correction controle valeur des cadrans
git-svn-id: https://svn.abls-habitat.fr/repo/Watchdog@5297 9f22af27-4b39-4cfe-9f85-4bed9fef4693

+++ trunk/Watchdogd/Serveur/cadran.c
@@ -63,7 +63,7 @@
           return( cadran->valeur != valeur );
         }
        case MNEMO_REGISTRE:
-        { gint valeur;
+        { gfloat valeur;
           valeur = Dls_data_get_R ( cadran->tech_id, cadran->acronyme, &cadran->dls_data );
           return( cadran->valeur != valeur );
         }
